finish implementing tests
