+++ packages/react-components/react-datepicker-compat/src/components/DatePicker/DatePicker.test.tsx
@@ -4,7 +4,6 @@ import { DatePicker } from './DatePicker';
 import { isConformant } from '../../testing/isConformant';
 import { datePickerClassNames } from './useDatePickerStyles';
 import { resetIdsForTests } from '@fluentui/react-utilities';
-import { calendarMonthClassNames } from '../CalendarMonth/useCalendarMonthStyles';
 
 // testing-library's queryByRole function doesn't look inside portals
 function queryByRoleDialog(result: RenderResult) {
@@ -66,6 +65,11 @@ describe('DatePicker', () => {
     expect(result.getByRole('combobox').getAttribute('readonly')).toBeNull();
   });
 
+  it('renders a readonly input when allowTextInput is false', () => {
+    const result = render(<DatePicker />);
+    expect(result.getByRole('combobox').getAttribute('readonly')).not.toBeNull();
+  });
+
   it('should call onSelectDate even when required input is empty when allowTextInput is true', () => {
     const onSelectDate = jest.fn();
     const result = render(<DatePicker isRequired allowTextInput onSelectDate={onSelectDate} />);
@@ -116,8 +120,8 @@ describe('DatePicker', () => {
   it('reflects the correct date in the input field when selecting a value and a different format is given', () => {
     const today = new Date('January 15, 2020');
     const initiallySelectedDate = new Date('January 10, 2020');
-    const onFormatDate = (date: Date): string => {
-      return date.getDate() + '/' + (date.getMonth() + 1) + '/' + (date.getFullYear() % 100);
+    const onFormatDate = (date?: Date): string => {
+      return date ? date.getDate() + '/' + (date.getMonth() + 1) + '/' + (date.getFullYear() % 100) : '';
     };
 
     const result = render(
@@ -134,58 +138,5 @@ describe('DatePicker', () => {
     result.getByText('15').click();
 
     expect(input.getAttribute('value')).toBe('15/1/20');
-  });
-});
-
-describe('When boundaries are specified', () => {
-  let result: RenderResult;
-  const defaultDate = new Date('Dec 15 2017');
-  const minDate = new Date('Jan 1 2017');
-  const maxDate = new Date('Dec 31 2017');
-  const strings = {
-    months: [
-      'January',
-      'February',
-      'March',
-      'April',
-      'May',
-      'June',
-      'July',
-      'August',
-      'September',
-      'October',
-      'November',
-      'December',
-    ],
-    shortMonths: ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'],
-    days: ['Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday'],
-    shortDays: ['S', 'M', 'T', 'W', 'T', 'F', 'S'],
-    goToToday: 'Go to today',
-    isOutOfBoundsErrorMessage: 'out of bounds',
-  };
-
-  beforeEach(() => {
-    result = render(
-      <DatePicker allowTextInput minDate={minDate} maxDate={maxDate} value={defaultDate} strings={strings} />,
-    );
-  });
-
-  afterEach(() => {
-    result.unmount();
-  });
-
-  it('should throw validation error for date outside boundary', () => {
-    const input = result.getByRole('combobox');
-    // before min date
-    fireEvent.change(input, { target: { value: 'Jan 1 2010' } });
-    fireEvent.blur(input);
-    expect(result.getByRole('alert')).toBeDefined();
-    expect(result.getByRole('alert').textContent).toBe('out of bounds');
-
-    // after max date
-    fireEvent.change(input, { target: { value: 'Jan 1 2020' } });
-    fireEvent.blur(input);
-    expect(result.getByRole('alert')).toBeDefined();
-    expect(result.getByRole('alert').textContent).toBe('out of bounds');
   });
 });
